Buttons to Showcase Game Files
I want it to be easy to see the game dev work I've done for anyone that clicks on this website, so having these buttons here will help. I'll upload said game dev work when I get the chance, in the meantime having the buttons redirect to a webpage that leads one to the game projects page showing the games I've contributed to.